refactor(bundler): reorganize modules
