+++ tooling/bundler/src/bundle/linux/mod.rs
@@ -1,0 +1,3 @@
+pub mod appimage;
+pub mod debian;
+pub mod rpm;
