+++ tooling/bundler/src/bundle/linux/templates/appimage
