+++ tooling/bundler/src/bundle/macos/templates/dmg/bundle_dmg
