+++ tooling/bundler/src/bundle/macos/templates/dmg/dmg-license.py
