+++ tooling/bundler/src/bundle/macos/templates/dmg/template.applescript
