+++ tooling/bundler/src/bundle/updater_bundle.rs
@@ -5,10 +5,10 @@
 use super::common;
 
 #[cfg(target_os = "macos")]
-use super::macos_bundle;
+use super::macos::app;
 
 #[cfg(target_os = "linux")]
-use super::appimage_bundle;
+use super::linux::appimage;
 
 #[cfg(target_os = "windows")]
 use super::msi_bundle;
@@ -56,7 +56,7 @@ fn bundle_update(settings: &Settings, bundles: &[Bundle]) -> crate::Result<Vec<P
         .find(|path| path.extension() == Some(OsStr::new("app")))
     }) {
     Some(path) => vec![path.clone()],
-    None => macos_bundle::bundle_project(settings)?,
+    None => app::bundle_project(settings)?,
   };
 
   // we expect our .app to be on bundle_path[0]
@@ -96,7 +96,7 @@ fn bundle_update(settings: &Settings, bundles: &[Bundle]) -> crate::Result<Vec<P
         .find(|path| path.extension() == Some(OsStr::new("AppImage")))
     }) {
     Some(path) => vec![path.clone()],
-    None => appimage_bundle::bundle_project(settings)?,
+    None => appimage::bundle_project(settings)?,
   };
 
   // we expect our .app to be on bundle[0]
