+++ tooling/bundler/src/bundle/windows/mod.rs
@@ -1,0 +1,1 @@
+pub mod msi;
