+++ tooling/bundler/src/bundle/windows/templates/main.wxs
